4089: Fix circular dependencies for shared (#4097)
* 4089: Fixed circular dependencies for shared

* 4089: Reverted getSlugFromPath and moved getCategoryTiles to categories.ts

* 4105: Add circular dependency checks for translations and build-configs (#4106)

* 4105: Add circular dependency checks for translations and build-configs

* 4105: Adjusted check-circular-dependencies at translations and buildConfigs
