+++ build-configs/package.json
@@ -16,6 +16,7 @@
   },
   "scripts": {
     "ts:check": "tsc --build",
-    "lint": "eslint --cache --cache-location ../.eslintcache ."
+    "lint": "eslint --cache --cache-location ../.eslintcache .",
+    "check-circular-dependencies": "yarn run -T check-circular-dependencies"
   }
 }
