+++ shared/api/endpoints/createCategoryChildrenEndpoint.ts
@@ -1,7 +1,8 @@
-import { Endpoint, EndpointBuilder, CategoryModel } from '..'
-
+import Endpoint from '../Endpoint'
+import EndpointBuilder from '../EndpointBuilder'
 import { API_VERSION } from '../constants'
 import mapCategoryJson from '../mapping/mapCategoryJson'
+import CategoryModel from '../models/CategoryModel'
 import { JsonCategoryType } from '../types'
 
 export const CATEGORY_CHILDREN_ENDPOINT_NAME = 'categoryChildren'
