+++ shared/index.ts
@@ -16,6 +16,7 @@ export * from './utils/replaceLinks'
 export * from './utils/normalizeString'
 export { default as findNormalizedMatches, findAllMatches } from './utils/findNormalizedMatches'
 export * from './utils'
+export { getCategoryTiles } from './utils/categories'
 export * from './constants/maps'
 export * from './constants/news'
 export * from './utils/dateFilterUtils'
