+++ shared/utils/isCurrentlyOpen.ts
@@ -1,6 +1,6 @@
 import { DateTime, Interval } from 'luxon'
 
-import { OpeningHoursModel } from '../api'
+import OpeningHoursModel from '../api/models/OpeningHoursModel'
 
 const isCurrentlyOpen = (openingHours: OpeningHoursModel[] | null): boolean => {
   if (!openingHours) {
